refactor: Move Apify error predicates out of logging into apify_errors
isActorRunLimitError, isMemoryQuotaError, and isPermissionApprovalError are
error classification, not logging — they only lived in logging.ts /
call_actor_common.ts for cycle reasons. Collect them (plus ACTOR_RUN_LIMIT_MESSAGE
and remoteMcpFailureDetail) in a dedicated leaf module utils/apify_errors.ts
that imports only const + apify-client, so logging, telemetry, payments, and the
tool layer share one source of truth with no import cycles.

logging.ts is logging-only again. buildCallActorErrorResponse now uses the
isPermissionApprovalError predicate instead of an inline duplicate check.
